% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{connect_afsc_akfin}
\alias{connect_afsc_akfin}
\title{Connect to AFSC and AKFIN databases}
\usage{
connect_afsc_akfin()
}
\value{
A list with AFSC and AKFIN DBI connections
}
\description{
Uses keyring to securely retrieve credentials.
}
