% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Connect to AFSC and AKFIN databases}
\usage{
db_connect()
}
\value{
A list with AFSC and AKFIN DBI connections
}
\description{
Uses keyring to securely retrieve credentials.
}
