% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_council_catch_data.R
\name{get_council_catch_data}
\alias{get_council_catch_data}
\title{Retrieve catch data by species and minimum year}
\usage{
get_council_catch_data(afsc, akfin, species, year_min)
}
\arguments{
\item{afsc}{DBI connection object afsc}

\item{akfin}{DBI connection object akfin}

\item{species}{Integer observer program species code}

\item{year_min}{Integer minimum year to include (inclusive)}

\item{sql_dir}{Directory containing SQL files (default "sql")}
}
\value{
A data.frame with catch data
}
\description{
Pulls catch data from the AFSC database using GET_ALL_CATCH.sql,
filtering by agency species code and year >= input year.
}
