% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_length_frequency_noaa.R
\name{plot_length_frequency_noaa}
\alias{plot_length_frequency_noaa}
\title{Plot length-frequency composition (with region & gear options)}
\usage{
plot_length_frequency_noaa(
  lf,
  species_name = NULL,
  date_min = NULL,
  date_max = NULL,
  region = NULL,
  areas = NULL,
  gear = c("Trawl", "Pot", "Longline"),
  facet_gear = FALSE,
  show_titles = TRUE,
  show_label = TRUE,
  show_n = TRUE,
  y_free = TRUE
)
}
\arguments{
\item{lf}{data.frame/data.table with length-frequency output}

\item{species_name}{Optional species label for title}

\item{date_min}{Optional start date (character "mm/dd/yyyy") for title only}

\item{date_max}{Optional end date (character "mm/dd/yyyy") for title only}

\item{region}{Optional region selector: "AI","BS","GOA","BSWGOA"}

\item{areas}{Optional explicit NMFS area codes (overrides region)}

\item{facet_gear}{If TRUE, facet by gear}

\item{show_titles}{If FALSE, remove title entirely}

\item{show_label}{If FALSE, remove upper-right label}

\item{show_n}{If TRUE, add sample-size label(s) in the upper-left}

\item{y_free}{If TRUE and faceting, allow free y-scales}
}
\value{
ggplot object
}
\description{
Expects \code{lf} with columns:
SPECIES, NMFS_AREA, GEAR, LENGTH, FREQUENCY
}
